requirements.txt
flask
requests
pyjwt
python-dotenv
Flask-SQLAlchemy
flask-jwt-extended
marshmallow
psycopg2